requirements.txt
loguru
pytest-playwright
fastapi[all]
uvicorn[standard]
.